refactor: 상수 클래스 제거하고 관련 도메인으로 이동
