--- src/main/java/model/GameRule.java
@@ -1,10 +1,0 @@
-package model;
-
-public final class GameRule {
-    private GameRule() {
-    }
-
-    public static final int DEALOUT_DRAW_COUNT = 2;
-    public static final int DEALER_DRAW_THRESHOLD = 16;
-    public static final int BLACKJACK_SCORE = 21;
-}
